+++ src/AuthenticationPlugin.php
@@ -2,6 +2,8 @@
 
 namespace Http\Client\Plugin;
 
+@trigger_error('The '.__NAMESPACE__.'\AuthenticationPlugin class is deprecated since version 1.1 and will be removed in 2.0. Use Http\Client\Common\Plugin\AuthenticationPlugin instead.', E_USER_DEPRECATED);
+
 use Http\Message\Authentication;
 use Psr\Http\Message\RequestInterface;
 
@@ -9,6 +11,8 @@ use Psr\Http\Message\RequestInterface;
  * Send an authenticated request.
  *
  * @author Joel Wurtz <joel.wurtz@gmail.com>
+ *
+ * @deprecated since since version 1.1, and will be removed in 2.0. Use {@link \Http\Client\Common\Plugin\AuthenticationPlugin} instead.
  */
 class AuthenticationPlugin implements Plugin
 {