+++ src/ErrorPlugin.php
@@ -2,6 +2,8 @@
 
 namespace Http\Client\Plugin;
 
+@trigger_error('The '.__NAMESPACE__.'\ErrorPlugin class is deprecated since version 1.1 and will be removed in 2.0. Use Http\Client\Common\Plugin\ErrorPlugin instead.', E_USER_DEPRECATED);
+
 use Http\Client\Plugin\Exception\ClientErrorException;
 use Http\Client\Plugin\Exception\ServerErrorException;
 use Psr\Http\Message\RequestInterface;
@@ -13,6 +15,8 @@ use Psr\Http\Message\ResponseInterface;
  * By default an exception will be thrown for all status codes from 400 to 599.
  *
  * @author Joel Wurtz <joel.wurtz@gmail.com>
+ *
+ * @deprecated since since version 1.1, and will be removed in 2.0. Use {@link \Http\Client\Common\Plugin\ErrorPlugin} instead.
  */
 class ErrorPlugin implements Plugin
 {