+++ src/HeaderDefaultsPlugin.php
@@ -2,6 +2,8 @@
 
 namespace Http\Client\Plugin;
 
+@trigger_error('The '.__NAMESPACE__.'\HeaderDefaultsPlugin class is deprecated since version 1.1 and will be removed in 2.0. Use Http\Client\Common\Plugin\HeaderDefaultsPlugin instead.', E_USER_DEPRECATED);
+
 use Psr\Http\Message\RequestInterface;
 
 /**
@@ -9,6 +11,8 @@ use Psr\Http\Message\RequestInterface;
  * If a given header already exists the value wont be replaced and the request wont be changed.
  *
  * @author Soufiane Ghzal <sghzal@gmail.com>
+ *
+ * @deprecated since since version 1.1, and will be removed in 2.0. Use {@link \Http\Client\Common\Plugin\HeaderDefaultsPlugin} instead.
  */
 class HeaderDefaultsPlugin implements Plugin
 {