+++ src/HeaderRemovePlugin.php
@@ -2,12 +2,16 @@
 
 namespace Http\Client\Plugin;
 
+@trigger_error('The '.__NAMESPACE__.'\HeaderRemovePlugin class is deprecated since version 1.1 and will be removed in 2.0. Use Http\Client\Common\Plugin\HeaderRemovePlugin instead.', E_USER_DEPRECATED);
+
 use Psr\Http\Message\RequestInterface;
 
 /**
  * Removes headers from the request.
  *
  * @author Soufiane Ghzal <sghzal@gmail.com>
+ *
+ * @deprecated since since version 1.1, and will be removed in 2.0. Use {@link \Http\Client\Common\Plugin\HeaderRemovePlugin} instead.
  */
 class HeaderRemovePlugin implements Plugin
 {