+++ src/HeaderSetPlugin.php
@@ -2,6 +2,8 @@
 
 namespace Http\Client\Plugin;
 
+@trigger_error('The '.__NAMESPACE__.'\HeaderSetPlugin class is deprecated since version 1.1 and will be removed in 2.0. Use Http\Client\Common\Plugin\HeaderSetPlugin instead.', E_USER_DEPRECATED);
+
 use Psr\Http\Message\RequestInterface;
 
 /**
@@ -9,6 +11,8 @@ use Psr\Http\Message\RequestInterface;
  * If the header does not exist it wil be set, if the header already exists it will be replaced.
  *
  * @author Soufiane Ghzal <sghzal@gmail.com>
+ *
+ * @deprecated since since version 1.1, and will be removed in 2.0. Use {@link \Http\Client\Common\Plugin\HeaderSetPlugin} instead.
  */
 class HeaderSetPlugin implements Plugin
 {