+++ src/LoggerPlugin.php
@@ -2,6 +2,8 @@
 
 namespace Http\Client\Plugin;
 
+@trigger_error('The '.__NAMESPACE__.'\LoggerPlugin class is deprecated since version 1.1 and will be removed in 2.0. Use Http\Client\Common\Plugin\LoggerPlugin instead.', E_USER_DEPRECATED);
+
 use Http\Client\Exception;
 use Http\Message\Formatter;
 use Http\Message\Formatter\SimpleFormatter;
@@ -13,6 +15,8 @@ use Psr\Log\LoggerInterface;
  * Log request, response and exception for a HTTP Client.
  *
  * @author Joel Wurtz <joel.wurtz@gmail.com>
+ *
+ * @deprecated since since version 1.1, and will be removed in 2.0. Use {@link \Http\Client\Common\Plugin\LoggerPlugin} instead.
  */
 class LoggerPlugin implements Plugin
 {